+++ src/app/modules/evaluate/provider/evaluate-query-item.provider.ts
@@ -55,6 +55,26 @@ export class EvaluateQueryItemProvider {
             }
         }
 
+        if (settings.evaluateQueryDefaultAttack) {
+            queryItem.damage = item.damage;
+
+            const prop = item.properties;
+            if (prop) {
+                queryItem.properties.weaponAttacksPerSecond = prop.weaponAttacksPerSecond;
+                queryItem.properties.weaponCriticalStrikeChance = prop.weaponCriticalStrikeChance;
+            }
+        }
+
+        if (settings.evaluateQueryDefaultDefense) {
+            const prop = item.properties;
+            if (prop) {
+                queryItem.properties.armourArmour = prop.armourArmour;
+                queryItem.properties.armourEvasionRating = prop.armourEvasionRating;
+                queryItem.properties.armourEnergyShield = prop.armourEnergyShield;
+                queryItem.properties.shieldBlockChance = prop.shieldBlockChance;
+            }
+        }
+
         if (!settings.evaluateQueryDefaultType) {
             if (item.rarity === ItemRarity.Normal ||
                 item.rarity === ItemRarity.Magic ||