+++ src/app/shared/module/poe/component/item-frame-value/item-frame-value.component.ts
@@ -26,10 +26,10 @@ export class ItemFrameValueComponent implements OnInit {
   public disabled: boolean;
 
   @Input()
-  public modifierMinRange: number;
+  public minRange: number;
 
   @Input()
-  public modifierMaxRange: number;
+  public maxRange: number;
 
   @Input()
   public value: ItemValue;
@@ -258,10 +258,10 @@ export class ItemFrameValueComponent implements OnInit {
     this.value.max = this.parsed;
     this.default = { ...this.value };
 
-    if (this.modifierMinRange === 0.5) {
-      this.value.min = this.value.tier?.min;
+    if (this.minRange === 0.5) {
+      this.value.min = undefined;
     } else {
-      this.value.min -= this.parsed * this.modifierMinRange;
+      this.value.min -= this.parsed * this.minRange;
       if (Number.isInteger(this.parsed)) {
         this.value.min = Math.floor(this.value.min);
       } else {
@@ -269,10 +269,10 @@ export class ItemFrameValueComponent implements OnInit {
       }
     }
 
-    if (this.modifierMaxRange === 0.5) {
-      this.value.max = this.value.tier?.max;
+    if (this.maxRange === 0.5) {
+      this.value.max = undefined;
     } else {
-      this.value.max += this.parsed * this.modifierMaxRange;
+      this.value.max += this.parsed * this.maxRange;
       if (Number.isInteger(this.parsed)) {
         this.value.max = Math.ceil(this.value.max);
       } else {